requirements.txt
feedparser==6.0.11
pyyaml==6.0.2
markdownify==0.12.1
